implemented onAuthExpired, onAuthRehydrated. refactored Luigi Core api and moved to own folder (so we can auto-generate the api afterwards)
