+++ core/src/Backdrop.html
@@ -1,7 +1,9 @@
-<div class="{backdropClass}" aria-hidden="false"><slot></slot></div>
+<div class="{backdropClass}" aria-hidden="false">
+  <slot></slot>
+</div>
 
 <script type="text/javascript">
-  import { LuigiConfig } from './services/config.js';
+  import { LuigiConfig } from './core-api';
   import * as iFrameHelpers from './utilities/helpers/iframe-helpers.js';
   export default {
     data() {
@@ -17,7 +19,7 @@
       if (!backdropDisabled) {
         this.setBackdropClass();
         window.addEventListener('message', e => {
-          if(!this.root.getValidMessageSource(e)) return;
+          if (!this.root.getValidMessageSource(e)) return;
           if ('luigi.add-backdrop' === e.data.msg) {
             this.set({ backdropActive: true });
           }
