+++ core/src/navigation/ContextSwitcher.html
@@ -1,3 +1,4 @@
+
 {#if contextSwitcherEnabled}
 <div class="fd-shellbar__action fd-shellbar__action--show-always">
   <div class="fd-product-menu">
@@ -34,7 +35,9 @@
           <ul class="fd-menu__list" id="context_menu_middle">
             {#if options && options.length === 0}
             <li>
-              <div class="fd-spinner"><div></div></div>
+              <div class="fd-spinner">
+                <div></div>
+              </div>
             </li>
             {/if} {#if options && options.length} {#each options as node}
             {#await getNodeName(node.label, config.fallbackLabelResolver,
@@ -43,12 +46,10 @@
               <a
                 class="fd-menu__item {node === selectedLabel ? 'is-selected' : ''}"
                 title="{label}"
-                >{label}</a
-              >
+              >{label}</a>
             </li>
             {/await} {/each} {/if}
           </ul>
-
           {#if actions && actions.length}
           <ul class="fd-menu__list fd-menu__list--bottom">
             {#each actions as node} {#if node.position === 'bottom'}
@@ -63,11 +64,11 @@
     </div>
   </div>
 </div>
-{/if}
 
+{/if}
 <script type="text/javascript">
   import { ContextSwitcherHelpers } from './services/context-switcher.js';
-  import { LuigiConfig } from '../services/config.js';
+  import { LuigiConfig } from '../core-api';
   import * as Helpers from '../utilities/helpers/generic-helpers.js';
   import * as Routing from '../services/routing.js';
   import * as RoutingHelpers from '../utilities/helpers/routing-helpers';
