+++ core/src/navigation/LeftNav.html
@@ -144,7 +144,7 @@
 <script type="text/javascript">
   import * as Navigation from './services/navigation.js';
   import * as NavigationHelpers from '../utilities/helpers/navigation-helpers';
-  import { LuigiConfig } from '../services/config.js';
+  import { LuigiConfig } from '../core-api';
 
   const setLeftNavData = async (current, component) => {
     const componentData = component.get();
