+++ core/src/navigation/LogoTitle.html
@@ -47,7 +47,7 @@
 {/if}
 <script type="text/javascript">
   import * as Header from './services/header.js';
-  import { LuigiConfig } from '../services/config.js';
+  import { LuigiConfig } from '../core-api';
   import { processHeaderSettings } from './services/header.js';
   import * as Routing from '../services/routing.js';
