+++ core/src/services/routing.js
@@ -2,7 +2,7 @@
 // Please consider adding any new methods to 'routing-helpers' if they don't require anything from this file.
 import * as Navigation from '../navigation/services/navigation';
 import * as RoutingHelpers from '../utilities/helpers/routing-helpers';
-import { LuigiConfig } from './config';
+import { LuigiConfig } from '../core-api';
 import * as GenericHelpers from '../utilities/helpers/generic-helpers';
 import * as Iframe from './iframe';
