+++ core/src/utilities/helpers/navigation-helpers.js
@@ -1,5 +1,5 @@
 // Helper methods for 'navigation.js' file. They don't require any method from 'navigation.js` but are required by them.
-import { LuigiConfig } from '../../services/config';
+import { LuigiConfig } from '../../core-api';
 import * as AuthHelpers from './auth-helpers';
 
 const EXP_CAT_KEY = 'luigi.preferences.navigation.expandedCategories';
