+++ core/test/services/iframe.spec.js
@@ -3,7 +3,7 @@ const assert = chai.assert;
 const sinon = require('sinon');
 const Iframe = require('../../src/services/iframe');
 import { afterEach } from 'mocha';
-import { LuigiConfig } from '../../src/services/config';
+import { LuigiConfig } from '../../src/core-api';
 
 describe('Iframe', () => {
   let node;
